+++ Changes.md
@@ -5,6 +5,9 @@ Changes in 2.0-alpha2
                 execute anything else than SELECTs on those databases then. All
                 buttons which usually let users edit or delete anything are not
                 displayed.
+- [FEATURE #3]  allow fed admins to upload a "minted" CA which will be auto-added
+                to new IdPs when they sign up. Good for federations where IdP
+                certificates come from one well-known CA.
 
 Configuration parameter changes
 -------------------------------